requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
pillow==10.4.0
qrcode==7.4.2
python-barcode==0.15.1
reportlab==4.2.0
httpx==0.27.0
pydantic[email]==2.9.0
python-dotenv==1.0.1
itsdangerous==2.2.0
aiosmtplib==3.0.1
email-validator==2.2.0
cryptography==42.0.0
redis==5.0.0
openpyxl==3.1.5
apscheduler>=3.10.0
